requirements.txt
streamlit
pypdf2
python-dotenv
nltk
tesseract
Pillow
pytesseract
deepseek-ai
google-generativeai
matplotlib
seaborn